refactor: modularize backend routes
