+++ backend/src/prisma.js
@@ -1,0 +1,5 @@
+const { PrismaClient, PostType } = require('@prisma/client');
+
+const prisma = new PrismaClient();
+
+module.exports = { prisma, PostType };
